Fixed percentage parsing in controller
I managed to remove the need for some hacks (isProxy, toRaw) from the
controller component. The main point seems to be to call functions
with the controller mode from the html side, rather than using the
precomputed constant in the defineComponent.

Still, can't get the unit to work with the same tricks so...